+++ test/snapshot/core/locals_assign/intended/main.go
@@ -4,7 +4,7 @@ import "snapshot/hxrt"
 
 func main() {
 	total := 1
-	total = int(int32((hxrt.Int32Wrap(total) + hxrt.Int32Wrap(4))))
+	total = int((hxrt.Int32Wrap(total) + hxrt.Int32Wrap(4)))
 	flag := false
 	flag = (total > 3)
 	hxrt.Println(any(total))
